+++ src/index.ts
@@ -3,15 +3,18 @@ import bunyan from 'bunyan';
 import logger from './logger';
 import AxiosLogger from './packages/AxiosLogger';
 import ExpressLogger from './packages/ExpressLogger';
+import RequestLogger from './packages/RequestLogger';
 import {
   IExpressLogger, IAxiosLogger, LoggerContext,
   LoggerConfig,
+  IRequestLogger,
 } from '../types';
 
 interface ILogger {
   AxiosLogger: IAxiosLogger;
   ExpressLogger: IExpressLogger;
   Logger: bunyan;
+  RequestLogger: IRequestLogger;
 }
 
 export default function init(config: LoggerConfig): ILogger {
@@ -25,5 +28,6 @@ export default function init(config: LoggerConfig): ILogger {
     Logger,
     AxiosLogger: new AxiosLogger(context),
     ExpressLogger: new ExpressLogger(context),
+    RequestLogger: new RequestLogger(context),
   };
 }