+++ types/index.d.ts
@@ -1,15 +1,7 @@
 import bunyan from 'bunyan';
-import { Request, Response, NextFunction } from 'express';
+import { Request as ERequest, Response, NextFunction } from 'express';
 import { AxiosStatic, AxiosInstance } from 'axios';
-
-/**
- * @property {string='logger'} PROJECT_NAME
- * @property {string[]=['/status', '/info']} OMIT_ROUTES
- */
-export interface Environment {
-  PROJECT_NAME: string;
-  OMIT_ROUTES: string[];
-}
+import { RequestAPI, CoreOptions, UriOptions, Request } from 'request';
 
 declare module 'express' {
   interface Request {
@@ -18,7 +10,7 @@ declare module 'express' {
 }
 
 export interface LoggerConfig {
-  PROJECT_NAME?: string;
+  PROJECT_NAME: string;
   OMIT_ROUTES?: string[];
 }
 
@@ -28,10 +20,14 @@ export interface LoggerContext {
 }
 
 export interface IExpressLogger {
-  onSuccess(req: Request, res: Response, next: NextFunction): void;
-  onError(error: Error, req: Request, res: Response, next: NextFunction): void;
+  onSuccess(req: ERequest, res: Response, next: NextFunction): void;
+  onError(error: Error, req: ERequest, res: Response, next: NextFunction): void;
 }
 
 export interface IAxiosLogger {
   attachInterceptor(axios: AxiosInstance): void;
+}
+
+export interface IRequestLogger {
+  attachDebug(requestPackage: RequestAPI<Request, CoreOptions, UriOptions>): void;
 }