+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
-
+### Added
+- Request Logger for request package
+- Description at README
 
 ## [1.1.0]
 ### Changed